Add files via upload
A clock on a green canvas. Short but simple. A Button to press to get the latest time. Good for a tkinter tutorial. Note: Has to be run in python 3. I'm using python 3.8, but I don't think the changes are significant enough between another version of python 3, but it's out there if you need it.